% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_gmail_smtp_creds}
\alias{create_gmail_smtp_creds}
\title{Create gmail smtp credentials}
\usage{
create_gmail_smtp_creds(user, format = c("key", "file"), overwrite = FALSE)
}
\arguments{
\item{user}{A string specifying gmail address}

\item{format}{A string value specifying credentials format key or file}

\item{overwrite}{A boolean value specifying overwrite or not only for key}
}
\description{
Create gmail smtp credentials
}
