% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_naver_smtp_creds}
\alias{create_naver_smtp_creds}
\title{Create naver smtp credentials}
\usage{
create_naver_smtp_creds(user, format = c("key", "file"), overwrite = FALSE)
}
\arguments{
\item{user}{A string specifying naver mail address}

\item{format}{A string value specifying credentials format key or file}

\item{overwrite}{A boolean value specifying overwrite or not}
}
\description{
Create naver smtp credentials
}
