% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send.R
\name{naver_send}
\alias{naver_send}
\title{Send from naver mail}
\usage{
naver_send(
  msg = "FYI",
  to,
  from,
  subject = NULL,
  cc = NULL,
  bcc = NULL,
  attach,
  creds = NULL,
  format = c("key", "file"),
  verbose = FALSE
)
}
\arguments{
\item{msg}{A text message}

\item{to}{A string vector specifying email address}

\item{from}{A string vector specifying email address}

\item{subject}{A string value specifying an email subject}

\item{cc}{A string vector specifying email address}

\item{bcc}{A string vector specifying email address}

\item{attach}{A string vector specifying attachment file paths}

\item{creds}{A credentials}

\item{format}{A string value specifying credentials format key or file}

\item{verbose}{A boolean value verbose output from the internal curl send_mail() call be printed or not}
}
\description{
\preformatted{
## Example
msg <- "For your infomation."
naver_send(msg = msg, to = "email_address")
}
}
