% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navergmail-package.R
\docType{package}
\name{navergmail-package}
\alias{navergmail}
\alias{navergmail-package}
\title{navergmail: Simple blastula package for naver mail and gmail}
\description{
This package is a simplified blastula package with default options for easy use of naver mail and gmail.
}
\details{
This package is created for easy use of blastula package with Naver mail and Gmail.
so this package contains a lot of default options (less flexibility).
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
